Fixed minor issues in English translation
When creating new alias, name and command titles should be in uppercase, but only one of them was. Command title does not need to have the asterisk, because the user can't add the alias without both fields filled, so I removed that. And I changed "Some lines of code has to be added..." to "Few lines of code will be added" so the user would know they have not been added yet.

BTW - great app, really aprreciate something like this!